dev-requirements.txt
backports.tarfile==1.2.0
black==25.1.0
build==1.2.2.post1
certifi==2025.6.15
cffi==1.17.1
cfgv==3.4.0
charset-normalizer==3.4.2
click==8.2.1
coverage==7.9.1
cryptography==45.0.4
distlib==0.3.9
docutils==0.21.2
filelock==3.18.0
flake8==7.3.0
id==1.5.0
identify==2.6.12
idna==3.10
importlib_metadata==8.7.0
iniconfig==2.1.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.2.1
keyring==25.6.0
markdown-it-py==3.0.0
mccabe==0.7.0
mdurl==0.1.2
more-itertools==10.7.0
msal==1.32.3
mypy_extensions==1.1.0
nh3==0.2.21
nodeenv==1.9.1
packaging==25.0
pathspec==0.12.1
platformdirs==4.3.8
pluggy==1.6.0
pre_commit==4.2.0
pycodestyle==2.14.0
pycparser==2.22
pyflakes==3.4.0
Pygments==2.19.2
PyJWT==2.10.1
pyproject_hooks==1.2.0
pytest==8.4.1
pytest-cov==6.2.1
pytest-mock==3.14.1
python-dotenv==1.0.1
PyYAML==6.0.2
readme_renderer==44.0
requests==2.32.4
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==14.0.0
twine==6.1.0
urllib3==2.5.0
virtualenv==20.31.2
zipp==3.23.0
